Clean up rsa code
